+++ Modules/Intent.Modules.Dapr.AspNetCore.Pubsub/release-notes.md
@@ -1,3 +1,7 @@
+### Version 2.0.0
+
+- Improvement: Updated to support use of top-level statements and minimal hosting model improvements introduced in [`Intent.AspNetCore` version 6.0.0](https://github.com/IntentArchitect/Intent.Modules.NET/blob/development/Modules/Intent.Modules.AspNetCore/release-notes.md#version-600).
+
 ### Version 1.1.1
 
 - Fixed: Nullability related compiler warnings.